Port the last of Create's main source set to 26.2

+++ src/main/java/com/simibubi/create/content/contraptions/elevator/ElevatorContactScreen.java
@@ -129,7 +129,7 @@ public class ElevatorContactScreen extends AbstractSimiScreen {
 
 		FormattedCharSequence formattedcharsequence = title.getVisualOrderText();
 		graphics.text(font, formattedcharsequence,
-			(float) (x + (background.getWidth() - 8) / 2 - font.width(formattedcharsequence) / 2), (float) y + 6, 0x2F3738, false);
+			x + (background.getWidth() - 8) / 2 - font.width(formattedcharsequence) / 2, y + 6, 0x2F3738, false);
 
 		GuiGameElement.of(AllBlocks.ELEVATOR_CONTACT.asStack()).<GuiGameElement
 				.GuiRenderBuilder>at(x + background.getWidth() + 6, y + background.getHeight() - 56, -200)
